Fix: return types changed from int to ssize_t for esp_serial_read/write

+++ host/linux/host_driver/esp32/esp_serial.c
@@ -46,7 +46,7 @@ static struct esp_serial_devs {
 	struct mutex lock;
 } devs[ESP_SERIAL_MINOR_MAX];
 
-static int esp_serial_read(struct file *file, char __user *user_buffer, size_t size, loff_t *offset)
+static ssize_t esp_serial_read(struct file *file, char __user *user_buffer, size_t size, loff_t *offset)
 {
 	struct esp_serial_devs *dev = NULL;
 	int ret_size = 0;
@@ -58,7 +58,7 @@ static int esp_serial_read(struct file *file, char __user *user_buffer, size_t s
 	return ret_size;
 }
 
-static int esp_serial_write(struct file *file, const char __user *user_buffer, size_t size, loff_t * offset)
+static ssize_t esp_serial_write(struct file *file, const char __user *user_buffer, size_t size, loff_t * offset)
 {
 	struct esp_payload_header *hdr = NULL;
 	u8 *tx_buf = NULL;
